Moved some of the conversion funcs around
